refactor: move TaskGroupingTreeStorage to TaskGroups
